captions/requirements.txt
mlx-whisper>=0.3.0
openai-whisper>=20231117
click>=8.1.0
lxml>=5.0.0
